change version id and move cruiseMethods in to seperate file
