+++ CruiseDAL/AutoGen/DAL.AutoGen.cs
@@ -3,7 +3,7 @@ namespace CruiseDAL
 {
     public partial class DAL
     {
-        internal const string CURENT_DBVERSION = "2.1.0";
+        internal const string CURENT_DBVERSION = "2.0.0";
         internal const int SCHEMA_VERSION = 2;
         internal const int MIN_COMPATIBLE_VERSION = 0;
     }
