+++ CruiseDAL/CruiseDAL_CF20.csproj
@@ -75,7 +75,7 @@
     <Compile Include="CruiseDAL\DataObject_Extra.cs" />
     <Compile Include="CruiseDAL\Enums.cs" />
     <Compile Include="CruiseDAL\Exceptions.cs" />
-    <Compile Include="CruiseDAL\Schema\Constants.cs" />
+    <Compile Include="CruiseDAL\Schema\CruiseMethods.cs" />
     <Compile Include="CruiseDAL\Schema\Schema.cs" />
     <Compile Include="CruiseDAL\Updater.cs" />
     <Compile Include="Logger\Logger.cs" />
