+++ CruiseDAL/Properties/AssemblyInfo_CF20.cs
@@ -31,5 +31,5 @@ using System.Runtime.InteropServices;
 //
 // You can specify all the values or you can default the Revision and Build Numbers 
 // by using the '*' as shown below:
-[assembly: AssemblyVersion("1.0.*")]
+[assembly: AssemblyVersion(CruiseDAL.DAL.CURENT_DBVERSION + ".*")]
