+++ CruiseDAL/TableModels/CruiseMethod_partial.cs
@@ -6,7 +6,7 @@ namespace CruiseDAL.DataObjects
         public static List<T> ReadCruiseMethods<T>(DAL db, bool reconMethodsOnly) where T : CruiseMethodsDO, new()
         {
             string condition = string.Format("WHERE Code NOT IN ( '{0}' ) {1}", 
-                string.Join("', '", CruiseDAL.Schema.Constants.CruiseMethods.UNSUPPORTED_METHODS),
+                string.Join("', '", CruiseDAL.Schema.CruiseMethods.UNSUPPORTED_METHODS),
                 (reconMethodsOnly) ? "AND Code = 'FIX' OR Code = 'PNT'" : string.Empty);
 
             return db.Read<T>(condition);
@@ -20,7 +20,7 @@ namespace CruiseDAL.DataObjects
         public static string[] ReadCruiseMethodStr(DAL db, bool reconMethodsOnly)
         {
             string command = string.Format("Select group_concat(Code,',') FROM CruiseMethods WHERE Code NOT IN ( '{0}' ) {1};",
-                string.Join("', '", CruiseDAL.Schema.Constants.CruiseMethods.UNSUPPORTED_METHODS),
+                string.Join("', '", CruiseDAL.Schema.CruiseMethods.UNSUPPORTED_METHODS),
                 (reconMethodsOnly) ? "AND Code = 'FIX' OR Code = 'PNT'" : string.Empty);
             string result = db.ExecuteScalar(command) as string ?? string.Empty;
             return result.Split(',');
